# Optimize UUID deserialization